+++ tests/unit/vpux_compiler/dialect/NPU50XX/regmapped_mapped_inference_test.cpp
@@ -305,7 +305,7 @@ TEST_F(NPUReg50XX_MappedInferenceTest, dpuPerfModeTest) {
             static_cast<nn_public::VpuHWPStatMode>(actual.read<vpux::NPUReg50XX::Fields::MiNNRTCfg_dpu_perf_mode>());
     EXPECT_EQ(actualValue, value);
 
-    reference.shv_rt_configs.dpu_perf_mode = value;
+    reference.shv_rt_configs.dpu_perf_mode = static_cast<uint8_t>(value);
     ASSERT_TRUE(isContentEqual());
 }
